added orchid admin panel, in process

+++ database/migrations/2023_05_27_204540_create_employees_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('employees', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->unsignedInteger('rank_id');
+            $table->unsignedBigInteger('rank_id');
             $table->string('photo');
             $table->string('full_rank')->nullable();
             $table->timestamps();
